+++ Backend/src/main/java/com/Backend/Repository/ComplaintRepo.java
@@ -4,7 +4,7 @@ import com.Backend.Entity.Complaints;
 import com.Backend.Entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ComplaintRepo extends JpaRepository<Complaints,Integer> {
-    Complaints findByComplaintId(int complaintId);
-    void deleteByComplaintId(int complaintId);
+public interface ComplaintRepo extends JpaRepository<Complaints,Long> {
+    Complaints findByComplaintId(long complaintId);
+    void deleteByComplaintId(long complaintId);
 }
